Fixed all gcc 4.4 warnings

+++ G3D.lib/source/BinaryInput.cpp
@@ -289,8 +289,8 @@ BinaryInput::BinaryInput(
             m_buffer = reinterpret_cast<uint8*>(System::alignedMalloc(m_length, 16));
             struct zip_file* zf = zip_fopen( z, internalFile.c_str(), ZIP_FL_NOCASE );
             {
-                int test = zip_fread( zf, m_buffer, m_length );
-                debugAssertM((size_t)test == m_length,
+                int64 test = zip_fread( zf, m_buffer, m_length );
+                debugAssertM(test == m_length,
                              internalFile + " was corrupt because it unzipped to the wrong size.");
                 (void)test;
             }
